requirements.txt
discord.py == 2.3.1
python-dotenv == 0.19.0